move tests from CodeListsV2ControllerSpec to CodeListsControllerSpec and update
