+++ conf/testOnlyDoNotUseInAppConf.routes
@@ -10,7 +10,7 @@
 # Failing to follow this rule may result in test routes deployed in production.
 
 # Add all the application routes to the prod.routes file
-->            /                                         app.v2.Routes
+->            /                                         app.Routes
 ->            /                                         definition.Routes
 
 GET           /crdl-cache/test-only/codelists             uk.gov.hmrc.crdlcache.controllers.testonly.TestOnlyController.codeListImportStatus()
