+++ it/test/uk/gov/hmrc/crdlcache/controllers/CodeListEntriesV2ControllerSpec.scala
@@ -77,7 +77,7 @@ class CodeListEntriesV2ControllerSpec
   }
 
   override def fakeApplication(): Application = GuiceApplicationBuilder()
-    .configure{"play.http.router" -> "app.v2.Routes"}
+    .configure{"play.http.router" -> "app.Routes"}
     .disable[InternalAuthModule]
     .overrides(
       bind[StubBehaviour].toInstance(authStub),
@@ -111,7 +111,7 @@ class CodeListEntriesV2ControllerSpec
         Json.obj("status" -> "active"),
         if withPhaseAndDomain then Some(entryPhase) else None,
         if withPhaseAndDomain then Some(entryDomain) else None
-      )  
+      )
     }
 
     private val expectedPDEntryPagedResult = (entries: Seq[CodeListEntry]) => Json.obj(
@@ -136,7 +136,7 @@ class CodeListEntriesV2ControllerSpec
         .thenReturn(Future.successful(5L))
 
       val response = httpClientV2
-        .get(url"http://localhost:$port/crdl-cache/v2/lists/$pdEntryCodeListCode/entries?pageNum=$defaultPageNum&pageSize=$defaultPageSize")
+        .get(url"http://localhost:$port/crdl-cache/admin/lists/$pdEntryCodeListCode/entries?pageNum=$defaultPageNum&pageSize=$defaultPageSize")
         .setHeader(HeaderNames.AUTHORIZATION -> "some-auth-token")
         .execute[HttpResponse]
         .futureValue
@@ -154,7 +154,7 @@ class CodeListEntriesV2ControllerSpec
         .thenReturn(Future.successful(5L))
 
       val response = httpClientV2
-        .get(url"http://localhost:$port/crdl-cache/v2/lists/$standardEntryCodeList/entries?pageNum=$defaultPageNum&pageSize=$defaultPageSize")
+        .get(url"http://localhost:$port/crdl-cache/admin/lists/$standardEntryCodeList/entries?pageNum=$defaultPageNum&pageSize=$defaultPageSize")
         .setHeader(HeaderNames.AUTHORIZATION -> "some-auth-token")
         .execute[HttpResponse]
         .futureValue
@@ -180,7 +180,7 @@ class CodeListEntriesV2ControllerSpec
         .thenReturn(Future.successful(5L))
 
       val response = httpClientV2
-        .get(url"http://localhost:$port/crdl-cache/v2/lists/$pdEntryCodeListCode/entries?pageNum=$defaultPageNum&pageSize=$defaultPageSize")
+        .get(url"http://localhost:$port/crdl-cache/admin/lists/$pdEntryCodeListCode/entries?pageNum=$defaultPageNum&pageSize=$defaultPageSize")
         .setHeader(HeaderNames.AUTHORIZATION -> "some-auth-token")
         .execute[HttpResponse]
         .futureValue
@@ -206,7 +206,7 @@ class CodeListEntriesV2ControllerSpec
         .thenReturn(Future.successful(5L))
 
       val response = httpClientV2
-        .get(url"http://localhost:$port/crdl-cache/v2/lists/$correspondenceEntryCodeList/entries?pageNum=$defaultPageNum&pageSize=$defaultPageSize")
+        .get(url"http://localhost:$port/crdl-cache/admin/lists/$correspondenceEntryCodeList/entries?pageNum=$defaultPageNum&pageSize=$defaultPageSize")
         .setHeader(HeaderNames.AUTHORIZATION -> "some-auth-token")
         .execute[HttpResponse]
         .futureValue
@@ -229,7 +229,7 @@ class CodeListEntriesV2ControllerSpec
       .thenReturn(Future.failed(new RuntimeException("Boom!!!")))
 
     val statusCode = try {
-        httpClientV2.get(url"http://localhost:$port/crdl-cache/v2/lists/$standardEntryCodeList/entries?pageNum=$defaultPageNum&pageSize=$defaultPageSize")
+        httpClientV2.get(url"http://localhost:$port/crdl-cache/admin/lists/$standardEntryCodeList/entries?pageNum=$defaultPageNum&pageSize=$defaultPageSize")
         .setHeader(HeaderNames.AUTHORIZATION -> "some-auth-token")
         .execute[HttpResponse]
         .futureValue
@@ -246,7 +246,7 @@ class CodeListEntriesV2ControllerSpec
 
     val (statusCode, responseJson) = try {
       val r = httpClientV2
-        .get(url"http://localhost:$port/crdl-cache/v2/lists/$standardEntryCodeList/entries?pageNum=$defaultPageNum&pageSize=$defaultPageSize")
+        .get(url"http://localhost:$port/crdl-cache/admin/lists/$standardEntryCodeList/entries?pageNum=$defaultPageNum&pageSize=$defaultPageSize")
         .execute[HttpResponse]
         .futureValue
       (r.status, r.json)
@@ -263,7 +263,7 @@ class CodeListEntriesV2ControllerSpec
 
     val (statusCode, responseJson) = try {
       val r = httpClientV2
-        .get(url"http://localhost:$port/crdl-cache/v2/lists/$standardEntryCodeList/entries?pageNum=$defaultPageNum&pageSize=$defaultPageSize")
+        .get(url"http://localhost:$port/crdl-cache/admin/lists/$standardEntryCodeList/entries?pageNum=$defaultPageNum&pageSize=$defaultPageSize")
         .setHeader(HeaderNames.AUTHORIZATION -> "some-auth-token")
         .execute[HttpResponse]
         .futureValue
@@ -282,7 +282,7 @@ class CodeListEntriesV2ControllerSpec
 
     val (statusCode, responseJson) = try {
       val r = httpClientV2
-        .get(url"http://localhost:$port/crdl-cache/v2/lists/$standardEntryCodeList/entries?pageNum=$defaultPageNum&pageSize=$defaultPageSize")
+        .get(url"http://localhost:$port/crdl-cache/admin/lists/$standardEntryCodeList/entries?pageNum=$defaultPageNum&pageSize=$defaultPageSize")
         .setHeader(HeaderNames.AUTHORIZATION -> "some-auth-token")
         .execute[HttpResponse]
         .futureValue
@@ -301,7 +301,7 @@ class CodeListEntriesV2ControllerSpec
 
     val statusCode = try {
       httpClientV2
-        .get(url"http://localhost:$port/crdl-cache/v2/lists/$standardEntryCodeList/entries?pageNum=$defaultPageNum&pageSize=$defaultPageSize")
+        .get(url"http://localhost:$port/crdl-cache/admin/lists/$standardEntryCodeList/entries?pageNum=$defaultPageNum&pageSize=$defaultPageSize")
         .setHeader(HeaderNames.AUTHORIZATION -> "some-auth-token")
         .execute[HttpResponse]
         .futureValue
