+++ it/test/uk/gov/hmrc/crdlcache/controllers/CodeListsControllerSpec.scala
@@ -30,7 +30,7 @@ import play.api.inject.bind
 import play.api.inject.guice.GuiceApplicationBuilder
 import play.api.libs.json.*
 import uk.gov.hmrc.crdlcache.controllers.auth.Permissions.ReadCodeLists
-import uk.gov.hmrc.crdlcache.models.CodeListCode.{BC08, BC36, BC66, CL231, E200}
+import uk.gov.hmrc.crdlcache.models.CodeListCode.{BC08, BC36, BC66, CL231, E200, CL251}
 import uk.gov.hmrc.crdlcache.models.{CodeListCode, CodeListEntry, LastUpdated}
 import uk.gov.hmrc.crdlcache.repositories.{CorrespondenceListsRepository, LastUpdatedRepository, StandardCodeListsRepository}
 import uk.gov.hmrc.http.HttpReads.Implicits.*
@@ -176,6 +176,18 @@ class CodeListsControllerSpec
         bind[Clock].toInstance(Clock.fixed(fixedInstant, ZoneOffset.UTC))
       )
       .build()
+
+  private val pagedResultJson = Json.obj(
+    "items" -> Json.arr(
+      Json.obj("codeListCode" -> "BC08", "snapshotVersion" -> 1, "lastUpdated" -> "2025-06-29T00:00:00Z"),
+      Json.obj("codeListCode" -> "BC66", "snapshotVersion" -> 1, "lastUpdated" -> "2025-06-28T00:00:00Z")
+    ),
+    "pageNum"     -> 1,
+    "pageSize"    -> 10,
+    "itemsInPage" -> 2,
+    "totalItems"  -> 2,
+    "totalPages"  -> 1
+  )
 
   "CodeListsController" should "return 200 OK when there are no errors" in {
     when(authStub.stubAuth(equalTo(Some(ReadCodeLists)), equalTo(Retrieval.EmptyRetrieval)))
@@ -794,5 +806,243 @@ class CodeListsControllerSpec
         .futureValue
 
     response.status mustBe Status.INTERNAL_SERVER_ERROR
+  }
+
+  "CodeListsController.fetchCodeListVersionsPaged" should "return 200 OK with a paged result when no filters are provided" in {
+    when(authStub.stubAuth(equalTo(Some(ReadCodeLists)), equalTo(Retrieval.EmptyRetrieval)))
+      .thenReturn(Future.unit)
+    when(lastUpdatedRepository.fetchAllLastUpdatedPaged(equalTo(1), equalTo(10), equalTo(None), equalTo(None), equalTo(None)))
+      .thenReturn(Future.successful(lastUpdatedEntries))
+    when(lastUpdatedRepository.codeListCount(equalTo(None), equalTo(None), equalTo(None)))
+      .thenReturn(Future.successful(2L))
+
+    val response = httpClientV2
+      .get(url"http://localhost:$port/crdl-cache/admin/lists?pageNum=1&pageSize=10")
+      .setHeader(HeaderNames.AUTHORIZATION -> "some-auth-token")
+      .execute[HttpResponse]
+      .futureValue
+
+    response.status mustBe Status.OK
+    response.json mustBe pagedResultJson
+  }
+
+  it should "return 200 OK with an empty items list when there are no matching code lists" in {
+    when(authStub.stubAuth(equalTo(Some(ReadCodeLists)), equalTo(Retrieval.EmptyRetrieval)))
+      .thenReturn(Future.unit)
+    when(lastUpdatedRepository.fetchAllLastUpdatedPaged(equalTo(1), equalTo(10), equalTo(None), equalTo(None), equalTo(None)))
+      .thenReturn(Future.successful(Seq.empty))
+    when(lastUpdatedRepository.codeListCount(equalTo(None), equalTo(None), equalTo(None)))
+      .thenReturn(Future.successful(0L))
+
+    val response = httpClientV2
+      .get(url"http://localhost:$port/crdl-cache/admin/lists?pageNum=1&pageSize=10")
+      .setHeader(HeaderNames.AUTHORIZATION -> "some-auth-token")
+      .execute[HttpResponse]
+      .futureValue
+
+    response.status mustBe Status.OK
+    response.json mustBe Json.obj(
+      "items" -> Json.arr(),
+      "pageNum" -> 1,
+      "pageSize" -> 10,
+      "itemsInPage" -> 0,
+      "totalItems" -> 0,
+      "totalPages" -> 0
+    )
+  }
+
+  it should "return 200 OK when a codeListCode filter is provided" in {
+    when(authStub.stubAuth(equalTo(Some(ReadCodeLists)), equalTo(Retrieval.EmptyRetrieval)))
+      .thenReturn(Future.unit)
+    when(lastUpdatedRepository.fetchAllLastUpdatedPaged(equalTo(1), equalTo(10), equalTo(Some("BC08")), equalTo(None), equalTo(None)))
+      .thenReturn(Future.successful(Seq(LastUpdated(BC08, 1, None, None, Instant.parse("2025-06-29T00:00:00Z")))))
+    when(lastUpdatedRepository.codeListCount(equalTo(Some("BC08")), equalTo(None), equalTo(None)))
+      .thenReturn(Future.successful(1L))
+
+    val response = httpClientV2
+      .get(url"http://localhost:$port/crdl-cache/admin/lists?pageNum=1&pageSize=10&codeListCode=BC08")
+      .setHeader(HeaderNames.AUTHORIZATION -> "some-auth-token")
+      .execute[HttpResponse]
+      .futureValue
+
+    response.status mustBe Status.OK
+  }
+
+  it should "return 200 OK when a phase filter is provided" in {
+    when(authStub.stubAuth(equalTo(Some(ReadCodeLists)), equalTo(Retrieval.EmptyRetrieval)))
+      .thenReturn(Future.unit)
+    when(lastUpdatedRepository.fetchAllLastUpdatedPaged(equalTo(1), equalTo(10), equalTo(None), equalTo(Some("P6")), equalTo(None)))
+      .thenReturn(Future.successful(Seq(LastUpdated(CL251, 1, Some("P6"), Some("NCTS"), Instant.parse("2025-06-29T00:00:00Z")))))
+    when(lastUpdatedRepository.codeListCount(equalTo(None), equalTo(Some("P6")), equalTo(None)))
+      .thenReturn(Future.successful(1L))
+
+    val response = httpClientV2
+      .get(url"http://localhost:$port/crdl-cache/admin/lists?pageNum=1&pageSize=10&phase=P6")
+      .setHeader(HeaderNames.AUTHORIZATION -> "some-auth-token")
+      .execute[HttpResponse]
+      .futureValue
+
+    response.status mustBe Status.OK
+  }
+
+  it should "return 200 OK when a domain filter is provided" in {
+    when(authStub.stubAuth(equalTo(Some(ReadCodeLists)), equalTo(Retrieval.EmptyRetrieval)))
+      .thenReturn(Future.unit)
+    when(lastUpdatedRepository.fetchAllLastUpdatedPaged(equalTo(1), equalTo(10), equalTo(None), equalTo(None), equalTo(Some("NCTS"))))
+      .thenReturn(Future.successful(Seq(LastUpdated(CL251, 1, Some("P6"), Some("NCTS"), Instant.parse("2025-06-29T00:00:00Z")))))
+    when(lastUpdatedRepository.codeListCount(equalTo(None), equalTo(None), equalTo(Some("NCTS"))))
+      .thenReturn(Future.successful(1L))
+
+    val response = httpClientV2
+      .get(url"http://localhost:$port/crdl-cache/admin/lists?pageNum=1&pageSize=10&domain=NCTS")
+      .setHeader(HeaderNames.AUTHORIZATION -> "some-auth-token")
+      .execute[HttpResponse]
+      .futureValue
+
+    response.status mustBe Status.OK
+  }
+
+  it should "return 200 OK when all three filters are provided" in {
+    when(authStub.stubAuth(equalTo(Some(ReadCodeLists)), equalTo(Retrieval.EmptyRetrieval)))
+      .thenReturn(Future.unit)
+    when(lastUpdatedRepository.fetchAllLastUpdatedPaged(equalTo(1), equalTo(10), equalTo(Some("CL251")), equalTo(Some("P6")), equalTo(Some("NCTS"))))
+      .thenReturn(Future.successful(Seq(LastUpdated(CL251, 1, Some("P6"), Some("NCTS"), Instant.parse("2025-06-29T00:00:00Z")))))
+    when(lastUpdatedRepository.codeListCount(equalTo(Some("CL251")), equalTo(Some("P6")), equalTo(Some("NCTS"))))
+      .thenReturn(Future.successful(1L))
+
+    val response = httpClientV2
+      .get(url"http://localhost:$port/crdl-cache/admin/lists?pageNum=1&pageSize=10&codeListCode=CL251&phase=P6&domain=NCTS")
+      .setHeader(HeaderNames.AUTHORIZATION -> "some-auth-token")
+      .execute[HttpResponse]
+      .futureValue
+
+    response.status mustBe Status.OK
+  }
+
+  it should "return 500 Internal Server Error when there is an error fetching from the repository" in {
+    when(authStub.stubAuth(equalTo(Some(ReadCodeLists)), equalTo(Retrieval.EmptyRetrieval)))
+      .thenReturn(Future.unit)
+    when(lastUpdatedRepository.fetchAllLastUpdatedPaged(equalTo(1), equalTo(10), equalTo(None), equalTo(None), equalTo(None)))
+      .thenReturn(Future.failed(new RuntimeException("Boom!!!")))
+    when(lastUpdatedRepository.codeListCount(equalTo(None), equalTo(None), equalTo(None)))
+      .thenReturn(Future.successful(0L))
+
+    val statusCode = try {
+      httpClientV2
+        .get(url"http://localhost:$port/crdl-cache/admin/lists?pageNum=1&pageSize=10")
+        .setHeader(HeaderNames.AUTHORIZATION -> "some-auth-token")
+        .execute[HttpResponse]
+        .futureValue
+        .status
+    } catch {
+      case e: UpstreamErrorResponse => e.statusCode
+    }
+
+    statusCode mustBe Status.INTERNAL_SERVER_ERROR
+  }
+
+  it should "return 401 Unauthorized when the user provides no Authorization header admin" in {
+    when(authStub.stubAuth(equalTo(Some(ReadCodeLists)), equalTo(Retrieval.EmptyRetrieval)))
+      .thenReturn(Future.unit)
+
+    val (statusCode, responseJson) = try {
+      val r = httpClientV2
+        .get(url"http://localhost:$port/crdl-cache/admin/lists?pageNum=1&pageSize=10")
+        .execute[HttpResponse]
+        .futureValue
+      (r.status, r.json)
+    } catch {
+      case e: UpstreamErrorResponse => (e.statusCode, Json.parse(e.message))
+    }
+
+    responseJson mustBe Json.obj("statusCode" -> 401, "message" -> "Unauthorized")
+    statusCode mustBe Status.UNAUTHORIZED
+  }
+
+  it should "return 401 Unauthorized when the user's token does not provide the appropriate permissions admin" in {
+    when(authStub.stubAuth(equalTo(Some(ReadCodeLists)), equalTo(Retrieval.EmptyRetrieval)))
+      .thenReturn(Future.failed(UpstreamErrorResponse("Unauthorized", Status.UNAUTHORIZED)))
+
+    val (statusCode, responseJson) = try {
+      val r = httpClientV2
+        .get(url"http://localhost:$port/crdl-cache/admin/lists?pageNum=1&pageSize=10")
+        .setHeader(HeaderNames.AUTHORIZATION -> "some-auth-token")
+        .execute[HttpResponse]
+        .futureValue
+      (r.status, r.json)
+    } catch {
+      case e: UpstreamErrorResponse => (e.statusCode, Json.parse(e.message))
+    }
+
+    responseJson mustBe Json.obj("statusCode" -> 401, "message" -> "Unauthorized")
+    statusCode mustBe Status.UNAUTHORIZED
+  }
+
+  it should "return 403 Forbidden when the user's token cannot be validated admin" in {
+    when(authStub.stubAuth(equalTo(Some(ReadCodeLists)), equalTo(Retrieval.EmptyRetrieval)))
+      .thenReturn(Future.failed(UpstreamErrorResponse("Forbidden", Status.FORBIDDEN)))
+
+    val (statusCode, responseJson) = try {
+      val r = httpClientV2
+        .get(url"http://localhost:$port/crdl-cache/admin/lists?pageNum=1&pageSize=10")
+        .setHeader(HeaderNames.AUTHORIZATION -> "some-auth-token")
+        .execute[HttpResponse]
+        .futureValue
+      (r.status, r.json)
+    } catch {
+      case e: UpstreamErrorResponse => (e.statusCode, Json.parse(e.message))
+    }
+
+    responseJson mustBe Json.obj("statusCode" -> 403, "message" -> "Forbidden")
+    statusCode mustBe Status.FORBIDDEN
+  }
+
+  it should "return 500 Internal Server Error when there is an error communicating with internal-auth admin" in {
+    when(authStub.stubAuth(equalTo(Some(ReadCodeLists)), equalTo(Retrieval.EmptyRetrieval)))
+      .thenReturn(Future.failed(UpstreamErrorResponse("Internal Server Error", 500)))
+
+    val statusCode = try {
+      httpClientV2
+        .get(url"http://localhost:$port/crdl-cache/admin/lists?pageNum=1&pageSize=10")
+        .setHeader(HeaderNames.AUTHORIZATION -> "some-auth-token")
+        .execute[HttpResponse]
+        .futureValue
+        .status
+    } catch {
+      case e: UpstreamErrorResponse => e.statusCode
+    }
+
+    statusCode mustBe Status.INTERNAL_SERVER_ERROR
+  }
+
+  "CodeListsController.getSnapShot" should "return 200 OK with a paged result when no filters are provided" in {
+    when(authStub.stubAuth(equalTo(Some(ReadCodeLists)), equalTo(Retrieval.EmptyRetrieval)))
+      .thenReturn(Future.unit)
+    when(lastUpdatedRepository.fetchLastUpdated(equalTo(BC08)))
+      .thenReturn(Future.successful(Some(LastUpdated(BC08, 1, None, None, Instant.parse("2025-06-29T00:00:00Z")))))
+
+    val response = httpClientV2
+      .get(url"http://localhost:$port/crdl-cache/admin/snapshot/BC08")
+      .setHeader(HeaderNames.AUTHORIZATION -> "some-auth-token")
+      .execute[HttpResponse]
+      .futureValue
+
+    response.status mustBe Status.OK
+    response.json mustBe Json.obj("codeListCode" -> "BC08", "snapshotVersion" -> 1, "lastUpdated" -> "2025-06-29T00:00:00Z")
+  }
+
+  "CodeListsController.getSnapShot" should "return 404 Not Found when no snapshot is found" in {
+    when(authStub.stubAuth(equalTo(Some(ReadCodeLists)), equalTo(Retrieval.EmptyRetrieval)))
+      .thenReturn(Future.unit)
+    when(lastUpdatedRepository.fetchLastUpdated(equalTo(BC08)))
+      .thenReturn(Future.successful(None))
+
+    val response = httpClientV2
+      .get(url"http://localhost:$port/crdl-cache/admin/snapshot/BC08")
+      .setHeader(HeaderNames.AUTHORIZATION -> "some-auth-token")
+      .execute[HttpResponse]
+      .futureValue
+
+    response.status mustBe Status.NOT_FOUND
   }
 }
